pyproject.toml
[tool.poetry]
name = "drink-water"
version = "0.1.0"
description = "System for register the quantity of water you drink"
authors = ["Bruno R. Scherer <brunoricardojava@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "~4.2"
djangorestframework = "~3.15.2"
pydantic = "~2.8.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.5.5"
black = "^24.4.2"
taskipy = "^1.13.0"
pytest = "^8.3.2"
pytest-django = "^4.8.0"
coverage = "^7.6.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = [
    "test_*.py",
    "*_test.py"
]

[tool.poetry.scripts]
format = "black ."
lint = "ruff ."

[tool.ruff]
target-version = "py311"
line-length = 120
exclude = ["core/", "manage.py", "migrations/", "*tests.py"]

[tool.black]
line-length = 120
target-version = ["py311"]
exclude = '''
/(
    \.venv
    | core
    | drink_water/migrations
    | .*tests\.py
)/
'''

[tool.taskipy.tasks]
format = "black ."
lint = "ruff check --fix ." 

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
